dlp_distributed/requirements.txt
asyncio==3.4.3
requests==2.32.3
pytest-asyncio==0.25.0
aiobotocore[boto3]==2.16.0
slack-sdk==3.34.0
django==5.1.4
pymysql==1.1.1